backend/requirements.txt
firebase-admin>=6.0.0
pyserial>=3.5
